Addbudgetpage still left with 1) error checking 2) maybe change threshold to slider 3) delete?

+++ Sanity/Category.h
@@ -11,7 +11,7 @@
 @interface Category : NSObject
 
 
-@property int limit;
+@property float limit;
 @property int total;
 @property (strong, atomic) NSString* name;
 @property (strong, atomic) NSString* budget;
